# Conflicts:
#	src/test/features/dashboard/routes/claimStatus/full-defence.ts

+++ src/main/features/response/routes/check-and-send.ts
@@ -20,10 +20,10 @@ import { ResponseDraft } from 'response/draft/responseDraft'
 import { Claim } from 'claims/models/claim'
 import { StatementOfMeansFeature } from 'response/helpers/statementOfMeansFeature'
 import { ClaimFeatureToggles } from 'utils/claimFeatureToggles'
-import { FeatureToggles } from 'utils/featureToggles'
 import { MediationDraft } from 'mediation/draft/mediationDraft'
 import { DirectionsQuestionnaireDraft } from 'directions-questionnaire/draft/directionsQuestionnaireDraft'
 import { FreeMediationUtil } from 'shared/utils/freeMediationUtil'
+import { FeatureToggles } from 'utils/featureToggles'
 
 const claimStoreClient: ClaimStoreClient = new ClaimStoreClient()
 
@@ -39,6 +39,7 @@ function renderView (form: Form<StatementOfTruth>, res: express.Response): void 
   }
   const statementOfTruthType = dqsEnabled ? SignatureType.DIRECTION_QUESTIONNAIRE : SignatureType.RESPONSE
   form.model.type = dqsEnabled ? SignatureType.DIRECTION_QUESTIONNAIRE : form.model.type
+  const mediationPilot: boolean = ClaimFeatureToggles.isFeatureEnabledOnClaim(claim, 'mediationPilot')
 
   res.render(Paths.checkAndSendPage.associatedView, {
     claim: claim,
@@ -47,14 +48,15 @@ function renderView (form: Form<StatementOfTruth>, res: express.Response): void 
     signatureType: signatureTypeFor(claim, draft),
     statementOfMeansIsApplicable: StatementOfMeansFeature.isApplicableFor(claim, draft.document),
     admissionsApplicable: ClaimFeatureToggles.isFeatureEnabledOnClaim(claim),
-    mediationEnabled: FeatureToggles.isEnabled('mediation'),
     dqsEnabled: dqsEnabled,
     mediationDraft: mediationDraft.document,
     contactPerson: FreeMediationUtil.getMediationContactPerson(claim, mediationDraft.document, draft.document),
     contactNumber: FreeMediationUtil.getMediationPhoneNumber(claim, mediationDraft.document, draft.document),
     directionsQuestionnaireDraft: directionsQuestionnaireDraft.document,
     datesUnavailable: datesUnavailable,
-    statementOfTruthType: statementOfTruthType
+    statementOfTruthType: statementOfTruthType,
+    mediationPilot: mediationPilot,
+    mediationEnabled: FeatureToggles.isEnabled('mediation')
   })
 }
 
@@ -148,11 +150,11 @@ export default express.Router()
         await claimStoreClient.saveResponseForUser(claim, draft, mediationDraft, directionsQuestionnaireDraft, user)
         await draftService.delete(draft.id, user.bearerToken)
 
-        if (FeatureToggles.isEnabled('mediation')) {
+        if (draft.document.response.type !== ResponseType.FULL_ADMISSION && mediationDraft.id) {
           await draftService.delete(mediationDraft.id, user.bearerToken)
         }
 
-        if (FeatureToggles.isEnabled('directionsQuestionnaire') && (draft.document.response.type === ResponseType.DEFENCE || draft.document.response.type === ResponseType.PART_ADMISSION)) {
+        if (FeatureToggles.isEnabled('directionsQuestionnaire') && directionsQuestionnaireDraft.id && (draft.document.response.type === ResponseType.DEFENCE || draft.document.response.type === ResponseType.PART_ADMISSION)) {
           await draftService.delete(directionsQuestionnaireDraft.id, user.bearerToken)
         }
